refactor(homePage): move homepage folder in /pages
