+++ packages/web-app/src/containers/SearchResultsContainer.jsx
@@ -3,7 +3,7 @@ import {
   fetchNextAdvancedsearchResults,
   fetchFullAdvancedsearchResults
 } from '../actions/Advancedsearch';
-import SearchResultsTable from '../components/homepage/advancedSearch/SearchResultsTable';
+import SearchResultsTable from '../pages/homepage/advancedSearch/SearchResultsTable';
 
 const getNewResults = (from, size) => dispatch => {
   dispatch(fetchNextAdvancedsearchResults(from, size));
