+++ packages/web-app/src/pages/homepage/DynamicNumber.jsx
