+++ packages/web-app/src/pages/homepage/Header.jsx
@@ -6,9 +6,9 @@ import {
   GridFullColumn
 } from '../../helpers/GridSystem';
 import { fseLinks } from '../../conf/Config';
-import InternationalizedLink from '../common/InternationalizedLink';
-import Translate from '../common/Translate';
-import GCLogo from '../common/GCLogo';
+import InternationalizedLink from '../../components/common/InternationalizedLink';
+import Translate from '../../components/common/Translate';
+import GCLogo from '../../components/common/GCLogo';
 
 const HeaderGridContainer = styled(GridContainer)`
   width: 100%;
