+++ packages/web-app/src/pages/homepage/PartnersCarousel.jsx
