+++ packages/web-app/src/pages/homepage/WhatIsIt/fonts/gc-icon.woff
