+++ packages/web-app/src/pages/homepage/index.jsx
@@ -2,18 +2,18 @@ import React, { useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 
-import { displayLoginDialog } from '../actions/Login';
+import { displayLoginDialog } from '../../actions/Login';
 
-import Header from '../components/homepage/Header';
-import Welcome from '../components/homepage/Welcome';
-import LatestBlogNewsSection from '../components/homepage/LatestBlogNewsSection';
-import Association from '../components/homepage/Association';
-import WhatIsIt from '../components/homepage/WhatIsIt';
-import RandomEntry from '../components/homepage/RandomEntry';
-import PartnersSection from '../components/homepage/PartnersSection';
-import Footer from '../components/homepage/Footer';
+import Header from './Header';
+import Welcome from './Welcome';
+import LatestBlogNewsSection from './LatestBlogNewsSection';
+import Association from './Association';
+import WhatIsIt from './WhatIsIt';
+import RandomEntry from './RandomEntry';
+import PartnersSection from './PartnersSection';
+import Footer from './Footer';
 
-import { usePermissions } from '../hooks';
+import { usePermissions } from '../../hooks';
 
 const HomePage = () => {
   const location = useLocation();
